Refactor account settings into multiple components, add qr code
