+++ package.json
@@ -21,6 +21,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
+    "react-qr-code": "^2.0.11",
     "react-router-dom": "^6.6.2",
     "typescript": "^4.9.4",
     "web-vitals": "^2.1.4"
